requirements.txt
simple-websocket
python-magic==0.4.24
eventlet==0.33.0
absl-py==0.15.0
gevent
flask_socketio
amqp==5.1.0
APScheduler==3.8.1
astunparse==1.6.3
bidict==0.21.4
bokeh==2.3.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
dnspython==2.2.1
Flask==2.0.1
Flask-Session==0.3.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pika==1.2.0
Pillow==8.4.0
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.5.4
seaborn==0.11.2
Shapely==1.8.1.post1
six==1.15.0
SQLAlchemy==1.4.31
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
thop==0.0.31.post2005241907
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.8
vine==5.0.0
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
